break change: remove ws port
- intergrate socket.io to remove wsport settings
- refactor some moduels
