+++ .env
@@ -1,7 +1,6 @@
 # public port
 MKIT_DATA=./data
 MKIT_PORT=8086
-MIKT_PORTWS=8087
 
 # qdrant ports
 QDRANT_PORT=6333
@@ -22,4 +21,3 @@ PSQL_PASS=justice
 # develop only
 #====================================
 IsDev=True
-HotReload=True
