+++ Dockerfile
@@ -5,10 +5,8 @@ FROM python:3.12
 WORKDIR /app
 
 ARG MKIT_PORT=8086
-ARG MKIT_PORTWS=8087
 
 ENV PORT=${MKIT_PORT}
-ENV PORTWS=${MKIT_PORTWS}
 
 RUN apt-get update && apt-get install -y curl && rm -rf /var/lib/apt/lists/*
 
@@ -19,7 +17,7 @@ RUN --mount=type=cache,target=/root/.cache \
 
 COPY src/ ./src/
 
-EXPOSE ${PORT} ${PORTWS}
+EXPOSE ${PORT}
 HEALTHCHECK CMD curl -f http://127.0.0.1:${PORT} || exit 1
 
 CMD ["python", "-m", "src.app"]
