+++ docker/different-host/.env
@@ -1,6 +1,5 @@
 # MediaKit settings
 MKIT_PORT=8086
-MKIT_PORTWS=8087
 MKIT_DATA=./data
