+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "immich-mediakit"
-version = "0.1.3"
+version = "0.1.4"
 
 [tool.pyright]
 include = ["src"]
