+++ requirements.txt
@@ -18,6 +18,7 @@ dash-table>=5.0.0,<6
 
 flask>=3.0.3,<3.1
 flask_caching>=2.3.1,<3
+flask-socketio~=5.5.1
 
 websockets>=15.0.1,<16
